Aula do dia 23/10/2024

+++ aulas/aula-07 (23-10-2024)/src/aula07/Listagem.java
@@ -19,7 +19,7 @@ public class Listagem {
     public static void main(String[] args) {
         int[] array1 = new int[] {7, 2, 9, 1, 10};
         listarArray(array1);
-        double[] array2 = new int[] {7.2, 2.7, 9.1, 1.9, 10.7};
+        double[] array2 = new double[] {7.2, 2.7, 9.1, 1.9, 10.7};
         listarArray(array2);
     }
